+++ src/main.ts
@@ -1,5 +1,6 @@
 import * as vsc from "vscode";
 import { EdifactFormattingEditProvider } from "./formattingProvider";
+import { putAllSegmentsInSingleLineCommandHandler } from "./commandHandlers";
 import { EdifactSymbolProvider } from "./symbolProvider";
 
 export function activate(ctx: vsc.ExtensionContext) {
@@ -10,5 +11,8 @@ export function activate(ctx: vsc.ExtensionContext) {
         vsc.languages.registerDocumentSymbolProvider(
             ["edifact"],
             new EdifactSymbolProvider()),
+        vsc.commands.registerCommand(
+            "edifact.putAllSegmentsToSingleLine",
+            putAllSegmentsInSingleLineCommandHandler)
     );
 }